fix: switch to a simpler test image.
The Flowers test image looked great, but it resulted in a really long
data-uri. I felt a little uncomfortable with that, given that these
files are designed to be copy/pasted.

Thus, I created a smaller and simpler test image, included it in the
repo, linked it from the README, and base64 encoded it into the html and
docs files. I think that fixes all the remaining broken images and
covers all the cases we needed.